app/requirements.txt
dash==2.9.3
dash_bootstrap_components==1.4.1
Flask_Caching==2.0.2
pandas==2.0.1
plotly==5.9.0
pymssql==2.2.5
gunicorn
dash-tools
pyarrow==12.0.0
fastparquet==0.5.0
lightgbm==3.3.5
python-dotenv==1.0.0